+++ src/main/java/org/kdb/inside/brains/view/inspector/model/NamespaceElement.java
@@ -1,6 +1,5 @@
 package org.kdb.inside.brains.view.inspector.model;
 
-import com.intellij.ide.util.treeView.smartTree.TreeElement;
 import icons.KdbIcons;
 
 import java.util.ArrayList;
@@ -15,19 +14,19 @@ public class NamespaceElement extends InspectorElement {
         this.item = item;
     }
 
-    protected static TreeElement[] buildChildren(NamespaceElement element, Object[] item) {
+    protected static InspectorElement[] buildChildren(NamespaceElement element, Object[] item) {
         final String namespace = element == null ? null : element.getCanonicalName();
 
-        final List<TreeElement> children = new ArrayList<>();
+        final List<InspectorElement> children = new ArrayList<>();
         Stream.of((Object[]) item[1]).forEach(s -> children.add(new FunctionElement(namespace, (Object[]) s)));
         Stream.of((Object[]) item[2]).forEach(s -> children.add(new TableElement(namespace, (Object[]) s)));
         Stream.of((Object[]) item[3]).forEach(s -> children.add(new VariableElement(namespace, (Object[]) s)));
         Stream.of((Object[]) item[4]).forEach(s -> children.add(new NamespaceElement(namespace, (Object[]) s)));
-        return children.toArray(TreeElement[]::new);
+        return children.toArray(InspectorElement[]::new);
     }
 
     @Override
-    protected TreeElement[] buildChildren() {
+    protected InspectorElement[] buildChildren() {
         return buildChildren(this, item);
     }
 }