+++ src/main/java/org/kdb/inside/brains/view/inspector/model/TableElement.java
@@ -12,7 +12,7 @@ public class TableElement extends ExecutableElement {
 
     public TableElement(String namespace, Object[] item) {
         super((String) item[0], namespace, KdbIcons.Node.Table);
-        size = (Long) item[1];
+        size = ((Number) item[1]).longValue();
         meta = (c.Flip) item[2];
         location = Array.getLength(meta.y[0]) + " columns, " + size + " rows, " + (isHistorical() ? "historical" : "memorable");
     }